+++ src/Resource/LocalResourceCollection.php
@@ -9,10 +9,9 @@
  * file that was distributed with this source code.
  */
 
-namespace Puli\Repository\Filesystem\Resource;
+namespace Puli\Repository\Resource;
 
 use Puli\Repository\Resource\Collection\ArrayResourceCollection;
-use Puli\Repository\Resource\Resource;
 
 /**
  * A collection of local resources.