+++ src/Resource/OverriddenPathLoader.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Puli\Repository\Filesystem\Resource;
+namespace Puli\Repository\Resource;
 
 /**
  * A loader for the overridden paths of a file.