+++ tests/FileCopyRepositoryTest.php
@@ -12,10 +12,9 @@
 namespace Puli\Repository\Tests;
 
 use Puli\Repository\FileCopyRepository;
-use Puli\Repository\Filesystem\Resource\LocalDirectoryResource;
-use Puli\Repository\Filesystem\Resource\LocalFileResource;
-use Puli\Repository\ManageableRepository;
 use Puli\Repository\Resource\DirectoryResource;
+use Puli\Repository\Resource\LocalDirectoryResource;
+use Puli\Repository\Resource\LocalFileResource;
 use Puli\Repository\ResourceRepository;
 use Symfony\Component\Filesystem\Filesystem;