+++ tests/FilesystemRepositoryTest.php
@@ -12,11 +12,11 @@
 namespace Puli\Repository\Tests;
 
 use Puli\Repository\FilesystemRepository;
-use Puli\Repository\Filesystem\Resource\LocalDirectoryResource;
-use Puli\Repository\Filesystem\Resource\LocalFileResource;
 use Puli\Repository\Resource\DirectoryResource;
 use Puli\Repository\Resource\Iterator\RecursiveResourceIteratorIterator;
 use Puli\Repository\Resource\Iterator\ResourceCollectionIterator;
+use Puli\Repository\Resource\LocalDirectoryResource;
+use Puli\Repository\Resource\LocalFileResource;
 use Puli\Repository\ResourceRepository;
 use Symfony\Component\Filesystem\Filesystem;