+++ tests/InMemoryRepositoryTest.php
@@ -12,7 +12,6 @@
 namespace Puli\Repository\Tests;
 
 use Puli\Repository\InMemoryRepository;
-use Puli\Repository\ManageableRepository;
 use Puli\Repository\Resource\Collection\ArrayResourceCollection;
 use Puli\Repository\Resource\DirectoryResource;
 use Puli\Repository\ResourceRepository;